[lib] metalRough reorg.
