+++ packages/lib/src/lib.ts
@@ -1,6 +1,6 @@
 export * from './ao';
 export * from './colorspace';
 export * from './dedup';
-export * from './metalrough';
+export * from './metal-rough';
 export * from './partition';
 export * from './inspect';
